+++ lib/RoadizUserBundle/src/State/UserValidationTokenProcessor.php
@@ -21,15 +21,15 @@ use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 final class UserValidationTokenProcessor implements ProcessorInterface
 {
     public function __construct(
-        private ManagerRegistry $managerRegistry,
-        private Roles $rolesBag,
-        private Security $security,
-        private EventDispatcherInterface $eventDispatcher,
-        private string $emailValidatedRoleName
+        private readonly ManagerRegistry $managerRegistry,
+        private readonly Roles $rolesBag,
+        private readonly Security $security,
+        private readonly EventDispatcherInterface $eventDispatcher,
+        private readonly string $emailValidatedRoleName
     ) {
     }
 
-    public function process($data, Operation $operation, array $uriVariables = [], array $context = []): VoidOutput
+    public function process(mixed $data, Operation $operation, array $uriVariables = [], array $context = []): VoidOutput
     {
         if (!$data instanceof UserValidationTokenInput) {
             throw new \RuntimeException(sprintf('Cannot process %s', get_class($data)));