+++ lib/Rozier/src/Controllers/AbstractAdminWithBulkController.php
@@ -84,13 +84,36 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
         if (null === $form) {
             return [];
         }
-        $ids = \json_decode($form->getData() ?? '[]');
+        if (!$form->isSubmitted() || !$form->isValid()) {
+            return [];
+        }
+        $json = $form->getData();
+        if (is_string($json)) {
+            $json = stripslashes(trim($json, '"'));
+        } else {
+            return [];
+        }
+        $ids = \json_decode($json, true);
+
         return \array_filter($ids, function ($id) {
             // Allow int or UUID identifiers
             return is_numeric($id) || is_string($id);
         });
     }
 
+    /**
+     * @param Request $request
+     * @param string $requiredRole
+     * @param FormInterface $bulkForm
+     * @param FormInterface $form
+     * @param callable(string): FormInterface $createBulkFormWithIds
+     * @param string $templatePath
+     * @param string $confirmMessageTemplate
+     * @param callable(PersistableInterface, FormInterface): void $alterItemCallable
+     * @param string $bulkFormName
+     * @return Response
+     * @throws \Twig\Error\RuntimeError
+     */
     protected function bulkAction(
         Request $request,
         string $requiredRole,
@@ -114,7 +137,7 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
                 $items = $this->getRepository()->findBy([
                     'id' => $ids,
                 ]);
-                $formWithIds = $createBulkFormWithIds(json_encode($ids));
+                $formWithIds = $createBulkFormWithIds(\json_encode($ids, JSON_THROW_ON_ERROR));
                 if (!$formWithIds instanceof FormInterface) {
                     throw new \RuntimeException('Invalid form returned.');
                 }
@@ -135,7 +158,11 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
                 ]);
                 foreach ($items as $item) {
                     if ($this->supports($item)) {
-                        $alterItemCallable($item);
+                        $alterItemCallable($item, $form);
+                        $updateEvent = $this->createUpdateEvent($item);
+                        if (null !== $updateEvent) {
+                            $this->dispatchSingleOrMultipleEvent($updateEvent);
+                        }
                         $msg = $this->getTranslator()->trans(
                             $confirmMessageTemplate,
                             [
@@ -147,7 +174,10 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
                     }
                 }
                 $this->em()->flush();
-                return $this->redirect($this->urlGenerator->generate($this->getDefaultRouteName()));
+                return $this->redirect($this->urlGenerator->generate(
+                    $this->getDefaultRouteName(),
+                    $this->getDefaultRouteParameters()
+                ));
             }
         }
 
@@ -163,6 +193,7 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
 
     public function bulkDeleteAction(Request $request): Response
     {
+        $this->additionalAssignation($request);
         return $this->bulkAction(
             $request,
             $this->getRequiredDeletionRole(),
@@ -184,6 +215,7 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
 
     public function bulkPublishAction(Request $request): Response
     {
+        $this->additionalAssignation($request);
         return $this->bulkAction(
             $request,
             $this->getRequiredRole(),
@@ -205,6 +237,7 @@ abstract class AbstractAdminWithBulkController extends AbstractAdminController
 
     public function bulkUnpublishAction(Request $request): Response
     {
+        $this->additionalAssignation($request);
         return $this->bulkAction(
             $request,
             $this->getRequiredRole(),