+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to Roadiz will be documented in this file.
 
+## [2.2.22](https://github.com/roadiz/core-bundle-dev-app/compare/v2.2.21...v2.2.22) - 2024-06-14
+
+### Bug Fixes
+
+- PHPStan errors fixes - ([7fe9bc5](https://github.com/roadiz/core-bundle-dev-app/commit/7fe9bc5c8580885922cdebf75797fa7d7773216d))
+- Pass FormInterface to `bulkAction` to update bulk item with a form field data. - ([a719cbe](https://github.com/roadiz/core-bundle-dev-app/commit/a719cbe6c07f4418888695b48bddd4075a3d80ab))
+
+## [2.2.21](https://github.com/roadiz/core-bundle-dev-app/compare/v2.2.20...v2.2.21) - 2024-06-13
+
+### Features
+
+- Added security layer on NodesSources form type for each node-type-field (NodeTypeFieldVoter) - ([cd3a962](https://github.com/roadiz/core-bundle-dev-app/commit/cd3a962b3f2df488ace2cc1d9709a117b11d96d7))
+- Added customizable `getDefaultRouteParameters` in AbstractAdminController - ([6f2e693](https://github.com/roadiz/core-bundle-dev-app/commit/6f2e693500cb97f32cf035a6e9073619cbbed191))
+
+## [2.2.20](https://github.com/roadiz/core-bundle-dev-app/compare/v2.2.19...v2.2.20) - 2024-06-06
+
+### Features
+
+- **(CoreBundle)** Made preview required RoleName configurable - ([3686d71](https://github.com/roadiz/core-bundle-dev-app/commit/3686d718b64ba12411c7b1c88262d401b95bfb9c))
+
 ## [2.2.19](https://github.com/roadiz/core-bundle-dev-app/compare/v2.2.18...v2.2.19) - 2024-06-04
 
 ### Features